+++ mobile/android/android-components/components/concept/storage/src/main/java/mozilla/components/concept/storage/HistoryMetadataStorage.kt
@@ -179,4 +179,9 @@ interface HistoryMetadataStorage {
      * Deletes metadata records that match [HistoryMetadataKey].
      */
     suspend fun deleteHistoryMetadata(key: HistoryMetadataKey)
+
+    /**
+     * Deletes metadata records that match [searchTerm] (case insensitive).
+     */
+    suspend fun deleteHistoryMetadata(searchTerm: String)
 }